Media sources added
Media sources are now added and working well, this makes it mush more usable as a dashboard.

I need to work on tweaking the UI elements appearance and position, but still looks nice atm I think, should be in next update.

+++ xbmc360/guilib/GUISpinControl.h
@@ -34,8 +34,8 @@ public:
 	void SetValue(int iValue);
 	int GetValue() const;
 
-	DWORD GetSpinWidth() const { return m_imgspinUp.GetWidth(); };
-	DWORD GetSpinHeight() const { return m_imgspinUp.GetHeight(); };
+	float GetSpinWidth() const { return m_imgspinUp.GetWidth(); };
+	float GetSpinHeight() const { return m_imgspinUp.GetHeight(); };
 
 	int GetMaximum() const;
 	int GetMinimum() const;
